Added IManagable interface to interfaces and fixed all new errors from it

+++ OrderApi/Repository/DeliveryTypeRepository.cs
@@ -1,4 +1,5 @@
 ﻿using Library.Extensions;
+using Library.Interfaces;
 using Microsoft.EntityFrameworkCore;
 using OrderApi.Data;
 using OrderApi.Models;
@@ -108,21 +109,15 @@ namespace OrderApi.Repository
             _logger.LogInformation($"Delivery type with Id[{deliveryType.DeliveryId}] updated succesfully.");
         }
 
-        public async Task DeleteAsync(DeliveryType deliveryType)
+        async public Task DeleteAsync(Guid id)
         {
-            try
-            {
-                _context.DeliveryTypes.Remove(deliveryType);
-                await _context.SaveChangesAsync();
-            }
-            catch (Exception ex)
-            {
-                _message = $"Deletion of Delivery type with id [{deliveryType.DeliveryId}] has failed.";
-                _logger.LogError(_message);
-                throw new ArgumentException(_message, ex);
-            }
+            var deliveryType = await _context.DeliveryTypes.FindAsync(id);
 
-            _logger.LogInformation($"Delivery type with Id [{deliveryType.DeliveryId}] deleted succesfully.");
+            if (deliveryType == null)
+                throw new KeyNotFoundException();
+
+            _context.DeliveryTypes.Remove(deliveryType);
+            await _context.SaveChangesAsync();
         }
     }
 }
